Confirm Delete/Cancel Task

+++ uk.co.compucorp.civicrm.tasksassignments/js/controllers/taskList.js
@@ -183,9 +183,30 @@ define(['controllers/controllers',
             };
 
             $scope.deleteTask = function(task){
-                 TaskService.delete(task.id).then(function(results){
-                     $scope.taskList.splice($scope.taskList.indexOf(task),1);
-                 });
+
+                var modalInstance = $modal.open({
+                    targetDomEl: $rootElement.find('div').eq(0),
+                    templateUrl: config.path.TPL+'modal/modalDialog.html',
+                    size: 'sm',
+                    controller: 'ModalDialogCtrl',
+                    resolve: {
+                        content: function(){
+                            return {
+                                msg: 'Are you sure you want to delete this task?'
+                            };
+                        }
+                    }
+                });
+
+                modalInstance.result.then(function(confirm){
+                    if (!confirm) {
+                        return
+                    }
+
+                    TaskService.delete(task.id).then(function(results){
+                        $scope.taskList.splice($scope.taskList.indexOf(task),1);
+                    });
+                });
             }
 
             $scope.$on('crmFormSuccess',function(e, data){
